+++ src/Moneybird.Net/Endpoints/Abstractions/IWorkflowEndpoint.cs
@@ -1,17 +1,9 @@
-using System.Collections.Generic;
-using System.Threading.Tasks;
+using Moneybird.Net.Endpoints.Abstractions.Common;
 using Moneybird.Net.Entities.Workflows;
 
 namespace Moneybird.Net.Endpoints.Abstractions
 {
-    public interface IWorkflowEndpoint
+    public interface IWorkflowEndpoint : IReadEndpoint<Workflow>
     {
-        /// <summary>
-        /// Get list of all the available workflows by access token.
-        /// </summary>
-        /// <param name="administrationId">The administration id.</param>
-        /// <param name="accessToken">The access token.</param>
-        /// <returns>A list of available workflow objects.</returns>
-        Task<List<Workflow>> GetWorkflowsAsync(string administrationId, string accessToken);
     }
 }