+++ src/Moneybird.Net/Endpoints/Abstractions/Options/IMoneybirdCreateOptions.cs
@@ -1,0 +1,6 @@
+namespace Moneybird.Net.Endpoints.Abstractions.Options
+{
+    public interface IMoneybirdCreateOptions
+    {
+    }
+}