+++ src/Moneybird.Net/Endpoints/SalesInvoices/Models/SalesInvoiceFilterOptions.cs
@@ -1,0 +1,25 @@
+using System;
+using Moneybird.Net.Endpoints.Abstractions.Options;
+using Moneybird.Net.Entities.SalesInvoices;
+
+namespace Moneybird.Net.Endpoints.SalesInvoices.Models
+{
+    public class SalesInvoiceFilterOptions : IMoneybirdFilterOptions
+    {
+        public SalesInvoiceState? State { get; set; }
+        
+        public string Period { get; set; }
+        
+        public string Reference { get; set; }
+
+        public int? ContactId { get; set; }
+        
+        public int? RecurringSalesInvoiceId { get; set; }
+        
+        public int? WorkflowId { get; set; }
+        
+        public DateTime? CreatedAfter { get; set; }
+        
+        public DateTime? UpdatedAfter { get; set; }
+    }
+}