+++ src/Moneybird.Net/Endpoints/SalesInvoices/Models/SalesInvoiceUpdateOptions.cs
@@ -1,0 +1,11 @@
+using System.Text.Json.Serialization;
+using Moneybird.Net.Endpoints.Abstractions.Options;
+
+namespace Moneybird.Net.Endpoints.SalesInvoices.Models
+{
+    public class SalesInvoiceUpdateOptions : IMoneybirdUpdateOptions
+    {
+        [JsonPropertyName("sales_invoice")]
+        public SalesInvoiceUpdate SalesInvoice { get; set; }
+    }
+}