+++ src/Moneybird.Net/Entities/SalesInvoices/SalesInvoiceCustomField.cs
@@ -1,0 +1,13 @@
+using System.Text.Json.Serialization;
+
+namespace Moneybird.Net.Entities.SalesInvoices
+{
+    public class SalesInvoiceCustomField
+    {
+        [JsonPropertyName("id")]
+        public int Id { get; set; }
+
+        [JsonPropertyName("value")]
+        public string Value { get; set; }
+    }
+}