+++ src/Moneybird.Net/Entities/Users/User.cs
@@ -1,10 +1,9 @@
 using System;
 using System.Text.Json.Serialization;
-using Moneybird.Net.Misc;
 
 namespace Moneybird.Net.Entities.Users
 {
-    public class User
+    public class User : IMoneybirdEntity
     {
         [JsonPropertyName("id")]
         public string Id { get; set; }