+++ src/Moneybird.Net/Entities/Workflows/Workflow.cs
@@ -1,10 +1,9 @@
 using System;
 using System.Text.Json.Serialization;
-using Moneybird.Net.Misc;
 
 namespace Moneybird.Net.Entities.Workflows
 {
-    public class Workflow
+    public class Workflow : IMoneybirdEntity
     {
         [JsonPropertyName("id")]
         public string Id { get; set; }