+++ src/Moneybird.Net/Http/Requester.cs
@@ -1,5 +1,5 @@
-using System;
 using System.Collections.Generic;
+using System.IO;
 using System.Net.Http;
 using System.Text;
 using System.Threading.Tasks;
@@ -13,14 +13,21 @@ namespace Moneybird.Net.Http
         {
         }
 
-        public async Task<string> CreateGetRequestAsync(string host, string relativeUrl, string accessToken, List<string> queryParameters = null)
+        public async Task<string> CreateGetRequestAsync(string host,
+            string relativeUrl,
+            string accessToken,
+            List<string> queryParameters = null)
         {
             var request = ConstructRequest(host, relativeUrl, accessToken, queryParameters, HttpMethod.Get);
             var response = await SendAsync(request).ConfigureAwait(false);
             return await GetResponseContentAsync(response).ConfigureAwait(false);
         }
 
-        public async Task<string> CreatePostRequestAsync(string host, string relativeUrl, string accessToken, string body, List<string> queryParameters = null)
+        public async Task<string> CreatePostRequestAsync(string host,
+            string relativeUrl,
+            string accessToken,
+            string body,
+            List<string> queryParameters = null)
         {
             var request = ConstructRequest(host, relativeUrl, accessToken, queryParameters, HttpMethod.Post);
             request.Content = new StringContent(body, Encoding.UTF8, "application/json");
@@ -29,7 +36,30 @@ namespace Moneybird.Net.Http
             return await GetResponseContentAsync(response).ConfigureAwait(false);
         }
 
-        public async Task<string> CreatePatchRequestAsync(string host, string relativeUrl, string accessToken, string body, List<string> queryParameters = null)
+        public async Task CreatePostFileRequestAsync(
+            string host,
+            string relativeUrl,
+            string accessToken,
+            string fileName,
+            Stream body,
+            List<string> queryParameters = null)
+        {
+            var request = ConstructRequest(host, relativeUrl, accessToken, queryParameters, HttpMethod.Post);
+            request.Content = new MultipartFormDataContent
+            {
+                { new StreamContent(body), "file", fileName }
+            };
+
+            var response = await SendAsync(request).ConfigureAwait(false);
+
+            response.EnsureSuccessStatusCode();
+        }
+
+        public async Task<string> CreatePatchRequestAsync(string host,
+            string relativeUrl,
+            string accessToken,
+            string body,
+            List<string> queryParameters = null)
         {
             var request = ConstructRequest(host, relativeUrl, accessToken, queryParameters, new HttpMethod("PATCH"));
             request.Content = new StringContent(body, Encoding.UTF8, "application/json");
@@ -38,7 +68,10 @@ namespace Moneybird.Net.Http
             return await GetResponseContentAsync(response).ConfigureAwait(false);
         }
 
-        public async Task<bool> CreateDeleteRequestAsync(string host, string relativeUrl, string accessToken, List<string> queryParameters = null)
+        public async Task<bool> CreateDeleteRequestAsync(string host,
+            string relativeUrl,
+            string accessToken,
+            List<string> queryParameters = null)
         {
             var request = ConstructRequest(host, relativeUrl, accessToken, queryParameters, HttpMethod.Delete);
             var response = await SendAsync(request).ConfigureAwait(false);