+++ src/Moneybird.Net/MoneybirdClient.cs
@@ -6,7 +6,11 @@ using Moneybird.Net.Endpoints.Administrations;
 using Moneybird.Net.Endpoints.Contacts;
 using Moneybird.Net.Endpoints.CustomFields;
 using Moneybird.Net.Endpoints.DocumentStyles;
+using Moneybird.Net.Endpoints.ExternalSalesInvoices;
+using Moneybird.Net.Endpoints.LegderAccounts;
 using Moneybird.Net.Endpoints.Payments;
+using Moneybird.Net.Endpoints.SalesInvoices;
+using Moneybird.Net.Endpoints.TaxRates;
 using Moneybird.Net.Endpoints.Users;
 using Moneybird.Net.Endpoints.Verifications;
 using Moneybird.Net.Endpoints.Workflows;
@@ -21,11 +25,16 @@ namespace Moneybird.Net
         private readonly Requester _requester;
 
         internal MoneybirdConfig Config { get; }
+        
         public IAdministrationEndpoint Administration { get; }
         public IContactEndpoint Contact { get; }
         public ICustomFieldEndpoint CustomField { get; }
         public IDocumentStyleEndpoint DocumentStyle { get; }
+        public IExternalSalesInvoiceEndpoint ExternalSalesInvoice { get; }
+        public ILedgerAccountEndpoint LedgerAccount { get; }
         public IPaymentEndpoint Payment { get; }
+        public ISalesInvoiceEndpoint SalesInvoice { get; }
+        public ITaxRateEndpoint TaxRate { get; }
         public IUserEndpoint User { get; }
         public IVerificationEndpoint Verification { get; }
         public IWorkflowEndpoint Workflow { get; }
@@ -60,7 +69,11 @@ namespace Moneybird.Net
             Contact = new ContactEndpoint(Config, _requester);
             CustomField = new CustomFieldEndpoint(Config, _requester);
             DocumentStyle = new DocumentStyleEndpoint(Config, _requester);
+            ExternalSalesInvoice = new ExternalSalesInvoiceEndpoint(Config, _requester);
+            LedgerAccount = new LedgerAccountEndpoint(Config, _requester);
             Payment = new PaymentEndpoint(Config, _requester);
+            SalesInvoice = new SalesInvoiceEndpoint(Config, _requester);
+            TaxRate = new TaxRateEndpoint(Config, _requester);
             User = new UserEndpoint(Config, _requester);
             Verification = new VerificationEndpoint(Config, _requester);
             Workflow = new WorkflowEndpoint(Config, _requester);