+++ tests/Moneybird.Net.Tests/Endpoints/Contacts/ContactEndpointTests.cs
@@ -46,7 +46,7 @@ namespace Moneybird.Net.Tests.Endpoints.Contacts
             var contactList = JsonSerializer.Deserialize<List<Contact>>(contactListJson, _config.SerializerOptions);
             Assert.NotNull(contactList);
 
-            var actualContacts = await _contactEndpoint.GetContactsAsync(AdministrationId, AccessToken);
+            var actualContacts = await _contactEndpoint.GetAsync(AdministrationId, AccessToken);
             Assert.NotNull(actualContacts);
 
             var actualContactList = actualContacts.ToList();
@@ -77,7 +77,7 @@ namespace Moneybird.Net.Tests.Endpoints.Contacts
                 LastName = "Doe"
             };
             
-            var actualContacts = await _contactEndpoint.GetContactsAsync(AdministrationId, AccessToken, contactFilterOptions);
+            var actualContacts = await _contactEndpoint.GetAsync(AdministrationId, AccessToken, contactFilterOptions);
             Assert.NotNull(actualContacts);
 
             var actualContactList = actualContacts.ToList();
@@ -194,7 +194,7 @@ namespace Moneybird.Net.Tests.Endpoints.Contacts
             var contact = JsonSerializer.Deserialize<Contact>(contactJson, _config.SerializerOptions);
             Assert.NotNull(contact);
 
-            var actualContact = await _contactEndpoint.GetContactByIdAsync(AdministrationId, ContactId, AccessToken);
+            var actualContact = await _contactEndpoint.GetByIdAsync(AdministrationId, ContactId, AccessToken);
             Assert.NotNull(actualContact);
 
             contact.Should().BeEquivalentTo(actualContact);
@@ -270,14 +270,14 @@ namespace Moneybird.Net.Tests.Endpoints.Contacts
             var contact = JsonSerializer.Deserialize<Contact>(contactJson, _config.SerializerOptions);
             Assert.NotNull(contact);
 
-            var actualContact = await _contactEndpoint.CreateContactAsync(AdministrationId, contactCreateOptions, AccessToken);
+            var actualContact = await _contactEndpoint.CreateAsync(AdministrationId, contactCreateOptions, AccessToken);
             Assert.NotNull(actualContact);
 
             contact.Should().BeEquivalentTo(actualContact);
         }
         
         [Fact]
-        public async void UpdateContactAsync_ByAccessToken_Returns_NewContact()
+        public async void UpdateContactAsync_ByAccessToken_Returns_UpdatedContact()
         {
             var contactJson = await File.ReadAllTextAsync(GetContactResponsePath);
             var contactUpdateOptions = new ContactUpdateOptions
@@ -315,7 +315,14 @@ namespace Moneybird.Net.Tests.Endpoints.Contacts
                     SiIdentifier = "",
                     SiIdentifierTypeType = null,
                     DirectDebit = false,
-                    CustomFieldsAttributes = null
+                    CustomFieldsAttributes = new List<ContactCustomFieldsAttribute>
+                    {
+                        new ()
+                        {
+                            Id = 1,
+                            Value = "Custom contact field"
+                        }
+                    }
                 }
             };
             
@@ -328,7 +335,7 @@ namespace Moneybird.Net.Tests.Endpoints.Contacts
             var contact = JsonSerializer.Deserialize<Contact>(contactJson, _config.SerializerOptions);
             Assert.NotNull(contact);
 
-            var actualContact = await _contactEndpoint.UpdateContactByIdAsync(AdministrationId, ContactId, contactUpdateOptions, AccessToken);
+            var actualContact = await _contactEndpoint.UpdateByIdAsync(AdministrationId, ContactId, contactUpdateOptions, AccessToken);
             Assert.NotNull(actualContact);
 
             contact.Should().BeEquivalentTo(actualContact);
@@ -340,7 +347,7 @@ namespace Moneybird.Net.Tests.Endpoints.Contacts
             _requester.Setup(moq => moq.CreateDeleteRequestAsync(It.IsAny<string>(), It.IsAny<string>(),
                 It.IsAny<string>(), It.IsAny<List<string>>())).ReturnsAsync(true);
             
-            var actualContact = await _contactEndpoint.DeleteContactByIdAsync(AdministrationId, ContactId, AccessToken);
+            var actualContact = await _contactEndpoint.DeleteByIdAsync(AdministrationId, ContactId, AccessToken);
             Assert.True(actualContact);
         }