+++ tests/Moneybird.Net.Tests/Endpoints/LedgerAccounts/LedgerAccountTestBase.cs
@@ -1,0 +1,6 @@
+namespace Moneybird.Net.Tests.Endpoints.LedgerAccounts;
+
+public class LedgerAccountTestBase : CommonTestBase
+{
+    protected static string LedgerAccountId = "387348630529377910";
+}